refactor(element/light-dom): move style utils to class as static property
