+++ packages/element/src/main.ts
@@ -1,8 +1,6 @@
 import {packageTracer} from '@alwatr/package-tracer';
 
-export * from './lib/base-element.js';
-export * from './lib/lit-light-dom-style.js';
+__dev_mode__: packageTracer.add(__package_name__, __package_version__);
+
 export * from './mixin/light-dom.js';
 export * from './mixin/logging.js';
-
-__dev_mode__: packageTracer.add(__package_name__, __package_version__);
